Removing auto-conversion of char32_t to utf-8.
Under Linux, everyone should be using UTF-8 consoles. However, that operation
is now deleted in C++20 so we cannot make use of it.

Further, at times we want to print the Unicode as an integer and if you
forget the cast you instead get the (possibly invalid) character.

+++ tools/show_unicode.cpp
@@ -141,7 +141,7 @@ int show_unicode::parse_args(int argc, char * argv[])
                 {
                     std::cerr
                         << "error: code \"0x"
-                        << std::uppercase << std::hex << std::setfill('0') << std::setw(6) << wc
+                        << std::uppercase << std::hex << std::setfill('0') << std::setw(6) << static_cast<std::uint32_t>(wc)
                         << "\" does not represent a valid Unicode character.\n";
                     return 1;
                 }
